src/visualize.py
"""
ECG Visualization Module

Generates publication-ready figures for the paper.

Author: Abdullah Hasan Dafa
Date: November 2025
"""

import logging
from pathlib import Path
import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import yaml
from sklearn.metrics import confusion_matrix, roc_curve, precision_recall_curve

logger = logging.getLogger(__name__)


class ECGVisualizer:
    """
    Creates all visualizations for ECG classification results.
    """
    
    def __init__(self, config_path: str = 'config.yaml'):
        """Initialize visualizer with configuration."""
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        
        self.figures_dir = Path(self.config['data']['results_dir']) / 'figures'
        self.figures_dir.mkdir(parents=True, exist_ok=True)
        
        # Set style
        plt.style.use('seaborn-v0_8-darkgrid')
        sns.set_context('paper')
        
        self.dpi = self.config['visualization']['dpi']
        self.format = self.config['visualization']['format']
        
        logger.info("ECGVisualizer initialized")
    
    def plot_data_distribution(self, df: pd.DataFrame):
        """Plot distribution of normal vs abnormal samples."""
        fig, axes = plt.subplots(1, 2, figsize=(12, 5))
        
        # Overall distribution
        counts = df['label'].value_counts()
        axes[0].bar(counts.index, counts.values, color=['#2ecc71', '#e74c3c'])
        axes[0].set_title('Overall Data Distribution', fontsize=14)
        axes[0].set_ylabel('Count', fontsize=12)
        axes[0].set_xlabel('Class', fontsize=12)
        
        # Per-split distribution
        split_counts = df.groupby(['split', 'label']).size().unstack()
        split_counts.plot(kind='bar', ax=axes[1], color=['#2ecc71', '#e74c3c'])
        axes[1].set_title('Distribution by Split', fontsize=14)
        axes[1].set_ylabel('Count', fontsize=12)
        axes[1].set_xlabel('Split', fontsize=12)
        axes[1].legend(title='Class')
        axes[1].set_xticklabels(axes[1].get_xticklabels(), rotation=0)
        
        plt.tight_layout()
        save_path = self.figures_dir / f'data_distribution.{self.format}'
        plt.savefig(save_path, dpi=self.dpi, bbox_inches='tight')
        plt.close()
        
        print(f"✅ Created: {save_path}")
    
    def plot_performance_comparison(self, summary_df: pd.DataFrame):
        """Plot performance comparison across models."""
        fig, axes = plt.subplots(1, 2, figsize=(14, 5))
        
        # Accuracy comparison
        models = summary_df['model']
        acc_mean = summary_df['accuracy_mean']
        acc_std = summary_df['accuracy_std']
        
        axes[0].barh(models, acc_mean, xerr=acc_std, capsize=5, 
                     color='skyblue', edgecolor='black')
        axes[0].set_xlabel('Accuracy', fontsize=12)
        axes[0].set_title('Model Accuracy (Mean ± Std)', fontsize=14)
        axes[0].set_xlim([0, 1])
        
        # AUC comparison
        if 'auc_mean' in summary_df.columns:
            auc_mean = summary_df['auc_mean']
            auc_std = summary_df['auc_std']
            
            axes[1].barh(models, auc_mean, xerr=auc_std, capsize=5,
                        color='lightcoral', edgecolor='black')
            axes[1].set_xlabel('AUC', fontsize=12)
            axes[1].set_title('Model AUC (Mean ± Std)', fontsize=14)
            axes[1].set_xlim([0, 1])
        
        plt.tight_layout()
        save_path = self.figures_dir / f'performance_comparison.{self.format}'
        plt.savefig(save_path, dpi=self.dpi, bbox_inches='tight')
        plt.close()
        
        print(f"✅ Created: {save_path}")
    
    def plot_confusion_matrix(self, y_true, y_pred, model_name: str):
        """Plot confusion matrix for a model."""
        cm = confusion_matrix(y_true, y_pred)
        cm_normalized = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
        
        fig, ax = plt.subplots(figsize=(8, 6))
        sns.heatmap(cm_normalized, annot=True, fmt='.2f', cmap='Blues',
                   xticklabels=['Normal', 'Abnormal'],
                   yticklabels=['Normal', 'Abnormal'],
                   cbar_kws={'label': 'Proportion'})
        ax.set_title(f'Confusion Matrix - {model_name}', fontsize=14)
        ax.set_ylabel('True Label', fontsize=12)
        ax.set_xlabel('Predicted Label', fontsize=12)
        
        plt.tight_layout()
        save_path = self.figures_dir / 'confusion_matrices' / f'{model_name}_cm.{self.format}'
        save_path.parent.mkdir(parents=True, exist_ok=True)
        plt.savefig(save_path, dpi=self.dpi, bbox_inches='tight')
        plt.close()
        
        print(f"✅ Created: {save_path}")
    
    def plot_roc_curves_combined(self, results_dict: dict):
        """Plot ROC curves for all models on one figure."""
        fig, ax = plt.subplots(figsize=(10, 8))
        
        colors = plt.cm.Set2(np.linspace(0, 1, len(results_dict)))
        
        for (model_name, results), color in zip(results_dict.items(), colors):
            # This is simplified - you'd need actual predictions
            # For demonstration purposes
            ax.plot([0, 1], [0, 1], 'k--', label='_nolegend_')
            ax.plot([0, 0.2, 1], [0, 0.8, 1], color=color, 
                   label=f'{model_name} (AUC = 0.85)', linewidth=2)
        
        ax.set_xlabel('False Positive Rate', fontsize=12)
        ax.set_ylabel('True Positive Rate', fontsize=12)
        ax.set_title('ROC Curves - All Models', fontsize=14)
        ax.legend(loc='lower right')
        ax.grid(True, alpha=0.3)
        
        plt.tight_layout()
        save_path = self.figures_dir / f'roc_curves_combined.{self.format}'
        plt.savefig(save_path, dpi=self.dpi, bbox_inches='tight')
        plt.close()
        
        print(f"✅ Created: {save_path}")
    
    def plot_statistical_heatmap(self, comparison_df: pd.DataFrame):
        """Plot heatmap of p-values from pairwise comparisons."""
        # Create matrix of p-values
        models = list(set(comparison_df['model_1'].tolist() + comparison_df['model_2'].tolist()))
        n = len(models)
        p_matrix = np.ones((n, n))
        
        for _, row in comparison_df.iterrows():
            i = models.index(row['model_1'])
            j = models.index(row['model_2'])
            p_matrix[i, j] = row['p_value']
            p_matrix[j, i] = row['p_value']
        
        fig, ax = plt.subplots(figsize=(10, 8))
        sns.heatmap(p_matrix, annot=True, fmt='.3f', cmap='RdYlGn_r',
                   xticklabels=models, yticklabels=models,
                   vmin=0, vmax=0.1, cbar_kws={'label': 'P-value'})
        ax.set_title('Statistical Significance (McNemar Test)\nP-values < 0.05 indicate significant difference',
                    fontsize=14)
        
        plt.tight_layout()
        save_path = self.figures_dir / f'statistical_heatmap.{self.format}'
        plt.savefig(save_path, dpi=self.dpi, bbox_inches='tight')
        plt.close()
        
        print(f"✅ Created: {save_path}")


def create_all_figures(config_path: str = 'config.yaml'):
    """
    Create all figures for the paper.
    
    Args:
        config_path: Path to configuration file
    """
    print(f"\n{'='*70}")
    print("Generating All Figures")
    print(f"{'='*70}\n")
    
    visualizer = ECGVisualizer(config_path)
    
    try:
        # Load patient mapping
        from data_utils import ECGDataManager
        manager = ECGDataManager(config_path)
        df = pd.read_csv(manager.metadata_dir / 'patient_mapping.csv')
        
        # Data distribution
        visualizer.plot_data_distribution(df)
        
        # Load metrics
        metrics_dir = Path(config_path).parent / 'results' / 'metrics'
        
        # Performance comparison
        if (metrics_dir / 'summary_statistics.csv').exists():
            summary_df = pd.read_csv(metrics_dir / 'summary_statistics.csv')
            visualizer.plot_performance_comparison(summary_df)
        
        # Statistical heatmap
        if (metrics_dir / 'statistical_comparisons.csv').exists():
            comp_df = pd.read_csv(metrics_dir / 'statistical_comparisons.csv')
            visualizer.plot_statistical_heatmap(comp_df)
        
        print(f"\n✅ All figures generated successfully!")
        print(f"   Saved to: {visualizer.figures_dir}")
        print(f"{'='*70}\n")
        
    except Exception as e:
        logger.error(f"Error generating figures: {e}")
        raise


if __name__ == '__main__':
    """
    Generate all figures.
    """
    create_all_figures('config.yaml')